+++ packages/paraglide/paraglide-js-adapter-astro/src/integration.ts
@@ -3,6 +3,7 @@ import { paraglide } from "@inlang/paraglide-js-adapter-vite"
 import path from "node:path"
 import { alias } from "./alias.js"
 import { fileURLToPath } from "node:url"
+import { normalizePath } from "./utilts.js"
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
@@ -23,6 +24,8 @@ export function integration(integrationConfig: {
 					entrypoint: middlewarePath,
 				})
 
+				const runtimePath = path.resolve(process.cwd(), integrationConfig.outdir, "runtime.js")
+
 				//Register the vite plugin
 				updateConfig({
 					vite: {
@@ -32,11 +35,10 @@ export function integration(integrationConfig: {
 								outdir: integrationConfig.outdir,
 							}),
 							alias({
-								"paraglide-js-adapter-astro:runtime": path.resolve(
-									process.cwd(),
-									integrationConfig.outdir,
-									"runtime.js"
-								),
+								//normalizing the path is very important!
+								//otherwise you get duplicate modules on windows
+								//learned that one the hard way (parjs-47)
+								"paraglide-js-adapter-astro:runtime": normalizePath(runtimePath),
 							}),
 						],
 					},