+++ packages/settings-component/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/settings-component
 
+## 0.0.11
+
+### Patch Changes
+
+- @inlang/sdk@0.28.2
+
+## 0.0.10
+
+### Patch Changes
+
+- @inlang/sdk@0.28.1
+
 ## 0.0.2
 
 ### Patch Changes